AEA-262: Split translation tests into separate files to reflect the code being tested.
